Add an easier-to-read way to delete a remote branch
Since 2010 Git supports the `--delete` argument for git-push as an
alias for the `:branch` syntax, i.e. these commands are equivalent:

    $ git push origin :foo
    $ git push origin --delete foo

This patch changes the tip to use the `--delete` argument on the
grounds that it is easier to remember, easier to understand, and
easily available since it is unlikely for people to be using a version
of Git which is more than five years old.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>