Describe how to see commits since forking from master
The command in this tip avoids showing merge commits because sometimes
developers will merge `master` into their feature branch in the course
of development, especially when that feature is in development for a
long time.  This tip focuses on showing only those commits which
introduce new changes and ignores those merge commits for this reason.
It also displays the commits in reverse chronological order, oldest to
newest, so that the output shows the development of a feature branch
in "historical" order.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>