Describe help for commonly-used Git commands with examples
This patch adds the command `git help everyday`, which shows a
man-page that comes with Git describing the most common commands for
everyday usage.  The tip is specifically at the top of the list on the
grounds that it is the fastest way to find help for performing basic
operations in Git, which is very useful to people new to Git.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>