Add a tip to show the most recent tag
This tip shows how to display the most recent tag reachable on the
current branch.  The use of `--abbrev=0` is necessary otherwise the
output would contain additional information (specifically part of a
commit hash), appended to the tag name.  It's worth noting that this
command is usable for branches other than the current one, e.g.

    $ git describe --tags --abbrev=0 master

That would show the most recent tag on the `master` branch regardless
of whatever branch the user is currently on.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>