Show how to read guides that come with Git
This patch adds a tip explaining `git help -g`.  That command will
list a collection of useful guides that come "pre-built" with Git,
e.g. `git help workflows` and `git help revisions`.  These are useful
resources for learning and looking-up common Git concepts and
techniques, and that broad usefulness is the justification for placing
the tip near the very top of the list.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>